Replaces deprecated add-path with appending to GITHUB_PATH.
Fixes #28

Fixes deprecation warning about "add-path".

Add Path was deprecated in
https://github.blog/changelog/2020-10-01-github-actions-deprecating-set-env-and-add-path-commands/
and will be removed soon, so this PR changes the mechanism to
the new one.

The path is added to the ${GITHUB_PATH} file instead as described in
https://docs.github.com/en/free-pro-team@latest/actions/reference/workflow-commands-for-github-actions#adding-a-system-path

It also adds lib folder to the repository, so that we can use
the action using commit hashes which is the recommended way of
using actions. See:

https://docs.github.com/en/free-pro-team@latest/actions/learn-github-actions/security-hardening-for-github-actions#using-third-party-actions